+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Add a Github Action for pushing to gh-pages on pushes to main
+- Add support for Mapbox Atlas.
 
 ## 0.12.0